wip: update list on demo and environment to use the same standards as tests and transactions
